feat: migrating to mantine wip
