+++ src/global.css
@@ -1,0 +1,6 @@
+html,
+body,
+#root {
+  width: 100vw;
+  height: 100vw;
+}
