+++ src/main.tsx
@@ -1,12 +1,17 @@
+// @ts-nocheck
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App.tsx";
-import "./index.css";
+import "@mantine/core/styles.css";
+import { MantineProvider } from "@mantine/core";
+import "./global.css";
 
 const root = document.getElementById("root");
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <MantineProvider defaultColorScheme="dark">
+      <App />
+    </MantineProvider>
   </React.StrictMode>,
   root
 );
